backend/requirements.txt
fastapi
uvicorn
crewai
langchain
anthropic
apify-client
exa-py
stripe
python-dotenv
pydantic
httpx
payments-py
